[IMRF-19] Switch to Hspec for running tests
Moved all logic from executable to library.
Made system tests call the library.
Made system tests use Hspec.
