+++ app/Options.hs
@@ -14,8 +14,8 @@ module Options
 import           Universum
 
 import           Options.Applicative (Parser, ParserInfo, argument, command, execParser,
-                                      fullDesc, help, helper, info, long, metavar,
-                                      progDesc, short, str, subparser, switch)
+                                      fullDesc, help, helper, info, metavar, progDesc,
+                                      str, subparser)
 
 data Command
     = SingleFile SingleFileOptions
@@ -24,7 +24,6 @@ data Command
 
 data SingleFileOptions = SingleFileOptions
     { sfoFilename    :: !FilePath -- ^ File to apply the tool to
-    , sfoPrintUnused :: !Bool     -- ^ Print unused Ids to stderr. For debug
     } deriving (Show)
 
 data CabalCacheOptions = CabalCacheOptions
@@ -44,11 +43,6 @@ fileParser = do
     sfoFilename <- argument str $
         metavar "FILE" <>
         help "File to importify"
-    sfoPrintUnused <- switch $
-        long "print-unused" <>
-        short 'U' <>
-        help "Print unused Ids to stderr. \
-             \Intended for debugging and testing purposes"
     pure $ SingleFile SingleFileOptions{..}
 
 cacheParser :: Parser Command
