+++ importify.cabal
@@ -20,8 +20,11 @@ library
                        Importify.CPP
                        Importify.Resolution
                        Importify.Tree
+                       Importify.Main
   build-depends:       base >= 4.7 && < 5
+                     , aeson
                      , Cabal
+                     , bytestring
                      , containers
                      , directory
                      , filepath
@@ -31,7 +34,9 @@ library
                      , hse-cpp
                      , path
                      , syb
+                     , text
                      , text-format
+                     , turtle
                      , universum
                      , unordered-containers
   default-language:    Haskell2010
@@ -52,18 +57,8 @@ executable importify
   main-is:             Main.hs
   other-modules:       Options
   build-depends:       base, importify
-                     , aeson
-                     , bytestring
-                     , containers
-                     , directory
                      , optparse-applicative
-                     , haskell-src-exts
-                     , haskell-names
-                     , path
-                     , text
-                     , turtle
                      , universum
-                     , unordered-containers
   default-language:    Haskell2010
 
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N -Wall
@@ -82,12 +77,12 @@ test-suite importify-system
   type:                exitcode-stdio-1.0
   main-is:             SystemSuite.hs
   hs-source-dirs:      test
-  test-module:         SystemSuite
   build-depends:       base, importify
                      , Cabal
                      , directory
-                     , haskell-src-exts
                      , filepath
+                     , haskell-src-exts
+                     , hspec
                      , text
                      , turtle
                      , universum
