+++ test/system/TestLongSpec.hs
@@ -1,8 +1,8 @@
--- (&&), (||), otherwise, bool
+-- &&, ||, otherwise, bool
 -- import Data.Bool (not)
 module TestLongSpec where
 
-import Data.Bool ((&&), (||), not, otherwise, bool)
+import Data.Bool (bool, not, otherwise, (&&), (||))
 
 func = not
